+++ models/evc.py
@@ -470,6 +470,54 @@ class EVCDeploy(EVCBase):
         self.sync()
         emit_event(self._controller, "undeployed", evc_id=self.id)
 
+    def remove_failover_flows(self, exclude_uni_switches=True,
+                              force=True, sync=True) -> None:
+        """Remove failover_flows.
+
+        By default, it'll exclude UNI switches, if mef_eline has already
+        called remove_current_flows before then this minimizes the number
+        of FlowMods and IO.
+        """
+        if not self.failover_path:
+            return
+        switches, cookie, excluded = OrderedDict(), self.get_cookie(), set()
+        links = set()
+        if exclude_uni_switches:
+            excluded.add(self.uni_a.interface.switch.id)
+            excluded.add(self.uni_z.interface.switch.id)
+        for link in self.failover_path:
+            if link.endpoint_a.switch.id not in excluded:
+                switches[link.endpoint_a.switch.id] = link.endpoint_a.switch
+                links.add(link)
+            if link.endpoint_b.switch.id not in excluded:
+                switches[link.endpoint_b.switch.id] = link.endpoint_b.switch
+                links.add(link)
+        for switch in switches.values():
+            try:
+                self._send_flow_mods(
+                    switch.id,
+                    [
+                        {
+                            "cookie": cookie,
+                            "cookie_mask": int(0xffffffffffffffff),
+                        }
+                    ],
+                    "delete",
+                    force=force,
+                )
+            except FlowModException as err:
+                log.error(
+                    f"Error removing flows from switch {switch.id} for"
+                    f"EVC {self}: {err}"
+                )
+        for link in links:
+            link.make_tag_available(link.get_metadata("s_vlan"))
+            link.remove_metadata("s_vlan")
+            notify_link_available_tags(self._controller, link)
+        self.failover_path = Path([])
+        if sync:
+            self.sync()
+
     def remove_current_flows(self, current_path=None, force=True):
         """Remove all flows from current path."""
         switches = set()
@@ -484,7 +532,7 @@ class EVCDeploy(EVCBase):
 
         match = {
             "cookie": self.get_cookie(),
-            "cookie_mask": 18446744073709551615,
+            "cookie_mask": int(0xffffffffffffffff)
         }
 
         for switch in switches:
@@ -515,7 +563,7 @@ class EVCDeploy(EVCBase):
                 dpid_flows_match[dpid].append({
                     "cookie": flow["cookie"],
                     "match": flow["match"],
-                    "cookie_mask": 18446744073709551615
+                    "cookie_mask": int(0xffffffffffffffff)
                 })
         for dpid, flows in self._prepare_uni_flows(path, skip_in=True).items():
             dpid_flows_match.setdefault(dpid, [])
@@ -608,7 +656,7 @@ class EVCDeploy(EVCBase):
                 use_path = Path()
                 self._install_direct_uni_flows()
             else:
-                log.warn(
+                log.warning(
                     f"{self} was not deployed. " "No available path was found."
                 )
                 return False
@@ -676,7 +724,7 @@ class EVCDeploy(EVCBase):
         self.sync()
 
         if not use_path:
-            log.warn(
+            log.warning(
                 f"Failover path for {self} was not deployed: {reason}"
             )
             return False
@@ -1004,11 +1052,11 @@ class EVCDeploy(EVCBase):
         """Check if current_path is deployed comparing with SDN traces."""
         trace_a = self.run_sdntrace(self.uni_a)
         if len(trace_a) != len(self.current_path) + 1:
-            log.warn(f"Invalid trace from uni_a: {trace_a}")
+            log.warning(f"Invalid trace from uni_a: {trace_a}")
             return False
         trace_z = self.run_sdntrace(self.uni_z)
         if len(trace_z) != len(self.current_path) + 1:
-            log.warn(f"Invalid trace from uni_z: {trace_z}")
+            log.warning(f"Invalid trace from uni_z: {trace_z}")
             return False
 
         for link, trace1, trace2 in zip(self.current_path,
@@ -1017,12 +1065,12 @@ class EVCDeploy(EVCBase):
             if compare_endpoint_trace(
                link.endpoint_a,
                glom(link.metadata, 's_vlan.value'), trace2) is False:
-                log.warn(f"Invalid trace from uni_a: {trace_a}")
+                log.warning(f"Invalid trace from uni_a: {trace_a}")
                 return False
             if compare_endpoint_trace(
                link.endpoint_b,
                glom(link.metadata, 's_vlan.value'), trace1) is False:
-                log.warn(f"Invalid trace from uni_z: {trace_z}")
+                log.warning(f"Invalid trace from uni_z: {trace_z}")
                 return False
 
         return True